chore: fix lint and update deps
