+++ package.json
@@ -26,19 +26,20 @@
     "android": "react-native run-android",
     "ios": "react-native run-ios",
     "storybook": "start-storybook -p 7007",
-    "storybook:web": "node ./node_modules/@storybook/react/bin/index.js start-storybook -p 9001 -c ./storybook-web",
+    "storybook:web": "node ./node_modules/@storybook/react/bin/index.js start-storybook -p 9001 -c storybook-web -s storybook-web/static",
     "cm": "git-cz",
     "release": "semantic-release",
     "build": "babel lib --out-dir dist --ignore node_modules,dist --source-maps --copy-files --no-babelrc --presets react-native",
     "build:sample:android": "react-native bundle --platform android --dev false --entry-file index.js --bundle-output android/app/src/main/assets/index.android.bundle --assets-dest android/app/src/main/res",
     "build:sample:ios": "react-native bundle --platform ios --dev false --entry-file index.js --bundle-output ios/NaturaReactComponents/main.jsbundle --assets-dest ios",
-    "build:storybook:web": "build-storybook -c ./storybook-web -o build/storybook",
+    "build:storybook:web": "build-storybook -c storybook-web -o build/storybook -s storybook-web/static",
     "archive:sample:ios": "xcodebuild -project ./ios/NaturaReactComponents.xcodeproj -scheme 'NaturaReactComponents' -derivedDataPath ./ios/build -archivePath ./ios/archive/NaturaReactComponents.xcarchive archive",
     "tsc": "tsc --noEmit",
     "postinstall": "npm dedupe && rm package-lock.json"
   },
   "dependencies": {
     "@naturacosmeticos/natds-icons": "^0.4.0",
+    "@naturacosmeticos/natds-styles": "^0.4.2",
     "prop-types": "15.7.2",
     "react-native-gesture-handler": "1.6.0",
     "react-native-svg": "11.0.1",
@@ -49,7 +50,6 @@
     "@babel/plugin-proposal-optional-chaining": "^7.9.0",
     "@babel/runtime": "7.6.2",
     "@naturacosmeticos/eslint-config-natura": "2.0.1",
-    "@naturacosmeticos/natds-styles": "^0.4.2",
     "@react-native-community/eslint-config": "0.0.5",
     "@storybook/addon-actions": "5.3.18",
     "@storybook/addon-docs": "^5.3.18",
@@ -57,7 +57,10 @@
     "@storybook/addon-links": "5.3.18",
     "@storybook/addon-viewport": "^5.3.18",
     "@storybook/addons": "5.3.18",
+    "@storybook/api": "^5.3.18",
     "@storybook/cli": "^5.3.18",
+    "@storybook/components": "^5.3.18",
+    "@storybook/core-events": "^5.3.18",
     "@storybook/react": "^5.3.18",
     "@storybook/react-native": "5.3.18",
     "@storybook/react-native-server": "5.3.18",
