+++ storybook-web/addons/theme/index.jsx
@@ -1,26 +1,32 @@
 import React, { useEffect, Fragment } from 'react';
 import { useAddonState, Consumer } from '@storybook/api';
 import { TabButton, WithTooltip, TooltipLinkList } from '@storybook/components';
-
-import { PANEL_ID, CHANGE, PARAM_KEY } from './shared';
 import { STORY_CHANGED } from '@storybook/core-events';
+import { PANEL_ID, CHANGE, PARAM_KEY } from './shared';
 
 import './styles.css';
 
 const DEFAULT_THEME = {
   name: 'natura',
-  type: 'light'
+  type: 'light',
 };
 
 function parseTheme(themes) {
   return Object.entries(themes).reduce((accum, [name, variants]) => {
-    Object.entries(variants).forEach(function ([type]) {
-      accum.push({ name, type })
-    })
+    Object.entries(variants).forEach(([type]) => {
+      accum.push({ name, type });
+    });
 
     return accum;
   }, []);
 }
+
+const createBackgroundSelectorItem = ({ name, type }, onChange, selectedItem) => ({
+  active: selectedItem.name === name && selectedItem.type === type,
+  id: `${name} - ${type}`,
+  onClick: () => onChange({ name, type }),
+  title: `${name} - ${type}`,
+});
 
 const getDisplayedItems = (list, onChange, selectedItem) => {
   if (!list.length) {
@@ -28,22 +34,13 @@ const getDisplayedItems = (list, onChange, selectedItem) => {
   }
 
   return [
-    ...list.map((params) =>
-      createBackgroundSelectorItem(params, onChange, selectedItem)
-    ),
+    ...list.map(params => createBackgroundSelectorItem(params, onChange, selectedItem)),
   ];
 };
 
-const createBackgroundSelectorItem = ({ name, type }, onChange, selectedItem) => ({
-  id: `${name} - ${type}`,
-  title: `${name} - ${type}`,
-  active: selectedItem.name === name && selectedItem.type === type,
-  onClick: () => onChange({ name, type })
-});
-
 const mapper = ({ api, state }) => {
   const story = state.storiesHash[state.storyId];
-  const { themes = {} } = story && api.getParameters(story.id, PARAM_KEY) || {};
+  const { themes = {} } = (story && api.getParameters(story.id, PARAM_KEY)) || {};
 
   return { items: parseTheme(themes) };
 };
@@ -51,40 +48,41 @@ const mapper = ({ api, state }) => {
 export default function Theme(props) {
   const { channel, api } = props;
   const [currentTheme, changeTheme] = useAddonState(PANEL_ID, DEFAULT_THEME);
-  const { disabled } = props.api.getCurrentParameter(PARAM_KEY) || {};
+  const { disabled } = api.getCurrentParameter(PARAM_KEY) || {};
 
   const handleChange = (params) => {
     changeTheme(params);
     channel.emit(CHANGE, params);
   };
 
+  const renderTooltip = ({ onHide }) => items => (
+    <TooltipLinkList
+      links={getDisplayedItems(items, (index) => {
+        handleChange(index);
+        onHide();
+      }, currentTheme)}
+    />
+  );
+
   useEffect(() => {
     channel.on(STORY_CHANGED, () => {
       changeTheme(DEFAULT_THEME);
     });
 
-    return () => (channel.removeListener(STORY_CHANGED))
+    return () => (channel.removeListener(STORY_CHANGED));
   }, []);
 
   if (disabled) return null;
 
   return (
     <Consumer filter={mapper}>
-      {({ items }) => {
-        return items.length ? (
+      {({ items }) => (items.length ? (
           <Fragment>
             <WithTooltip
               placement="top"
               trigger="click"
               closeOnClick
-              tooltip={({ onHide }) => (
-                <TooltipLinkList
-                  links={getDisplayedItems(items, i => {
-                    handleChange(i);
-                    onHide();
-                  }, currentTheme)}
-                />
-              )}
+              tooltip={renderTooltip(items)}
             >
               <TabButton
                 key="Theme"
@@ -93,8 +91,7 @@ export default function Theme(props) {
               </TabButton>
             </WithTooltip>
           </Fragment>
-        ) : null;
-      }}
+      ) : null)}
     </Consumer>
   );
 }
