+++ storybook-web/manager-head.html
@@ -3,6 +3,7 @@
   .sidebar-container .sidebar-item:hover {
     color: #f0f0f0;
   }
+
   .sidebar-container .sidebar-item.selected {
     background: rgba(255, 107, 11, 0.16);
   }
@@ -17,13 +18,15 @@
     box-shadow: #fff 0 0 0 1px inset;
   }
 
-  .sidebar-header img {
-    max-width: 80%;
-    margin-left: auto;
+  .sidebar-header a:after {
+    content: 'DS React Native';
+    color: white;
+    display: block;
+    text-align: center;
   }
 
   .simplebar-content .sidebar-subheading {
     color: #fff;
   }
 </style>
-<link rel="shortcut icon" href="images/favicon.png" />
+<link rel="shortcut icon" href="/static/favicon.png" />
