+++ storybook-web/theme.js
@@ -1,25 +1,18 @@
 import { create } from '@storybook/theming/create';
 
-// require('../src/assets/images/favicon.png');
-
 export default create({
-  base: 'light',
-
   appBg: '#333',
-  appContentBg: '#fff',
   appBorderRadius: 0,
-
-  colorPrimary: '#F49B00',
-  colorSecondary: '#EF8426',
-
-  barTextColor: '#f0f0f0',
-  barSelectedColor: '#F49B00',
+  appContentBg: '#fff',
   barBg: '#333',
-
-  textColor: '#333',
-  textInverseColor: '#fff',
-
+  barSelectedColor: '#F49B00',
+  barTextColor: '#f0f0f0',
+  base: 'light',
+  brandImage: '/logo-natura-negative-vertical.svg',
   brandTitle: 'natds-js',
   brandUrl: '/',
-  // brandImage: require('../src/assets/images/logo-natura.png'),
+  colorPrimary: '#F49B00',
+  colorSecondary: '#EF8426',
+  textColor: '#333',
+  textInverseColor: '#fff',
 });
